manually derive Hashable for recursive data types
Previous issues with "simplifier ticks exhausted" are due to the
derivations of `Hashable` in `Hashable.hs`, due to a ton of GHC
inlining. Manaully deriving the hashes of some of the largest types
reduces simplifier ticks and compilation time.

On my machine, compilation times go from 12+GB and a few hours to 3GB
and minutes.